+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * add php8 version support
 * change in scope «CRM» Product service and integration tests
 * add `AddedItemIdResultInterface` for batch-queries result
+* fix method name in `ContactsResult`
 
 ## 2.0-alpha.2 (31.01.2021)